+++ src/main/scala/contract/service/todo/AddItemService.scala
@@ -3,10 +3,10 @@ package contract.service.todo
 import contract.service.Service
 import domain.todo.Item
 
-abstract class AddItemService extends Service[AddItemService.Request, Map[Int, Item]]
+abstract class AddItemService extends Service[AddItemService.Request, Map[Long, Item]]
 
 object AddItemService {
 
-  case class Request(userId: Int, body: String, state: Boolean)
+  case class Request(userId: Long, body: String, state: Boolean)
 
 }
