+++ src/main/scala/contract/service/todo/EditStateService.scala
@@ -4,10 +4,10 @@ import contract.service.Service
 import domain.todo.Item
 
 
-abstract class EditStateService extends Service[EditStateService.Request, Map[Int, Item]]
+abstract class EditStateService extends Service[EditStateService.Request, Map[Long, Item]]
 
 object EditStateService {
 
-  case class Request(userId: Int, id: Int, newState: Boolean)
+  case class Request(userId: Long, id: Long, newState: Boolean)
 
 }
