+++ src/main/scala/domain/auth/Session.scala
@@ -1,6 +1,6 @@
 package domain.auth
 
-case class Session(userId: Int, isLogin: Boolean) {
+case class Session(userId: Long, isLogin: Boolean) {
 
   def setLogin(login: Boolean): Session = {
     copy(isLogin = login)
